+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add `MollieEx.ClientLinks.create/3`.
 - Add `MollieEx.Balances.list/2`, `MollieEx.Balances.get/3`, and
   `MollieEx.Balances.primary/2`.
+- Add `MollieEx.Balances.list_transactions/3`.
 
 ## 0.4.0 - 2026-06-14